// agent/src/types/vapi.types.ts
export enum VapiWebhookEnum {
    ASSISTANT_REQUEST = "assistant-request",
    FUNCTION_CALL = "function-call",
    STATUS_UPDATE = "status-update",
    END_OF_CALL_REPORT = "end-of-call-report",
    HANG = "hang",
    SPEECH_UPDATE = "speech-update",
    TRANSCRIPT = "transcript",
}

export interface ConversationMessage {
    role: "user" | "system" | "bot" | "function_call" | "function_result";
    message?: string;
    name?: string;
    args?: string;
    result?: string;
    time: number;
    endTime?: number;
    secondsFromStart: number;
}

interface BaseVapiPayload {
    call: VapiCall;
}
// ...
export interface EndOfCallReportPayload extends BaseVapiPayload {
    type: VapiWebhookEnum.END_OF_CALL_REPORT;
    endedReason: string;
    transcript: string;
    messages: ConversationMessage[];
    summary: string;
    recordingUrl?: string;
    call: {
        id: string;
        customer: {
            number: string;
        }
    }
}
// ...
export interface VapiCall {
}